requirements.txt
# Build tools (needed for Python 3.13)
setuptools>=69.0.0
wheel>=0.42.0

# Core Dependencies - Updated for Python 3.13 compatibility
numpy>=1.26.0
pandas>=2.1.0
scikit-learn>=1.3.2
xgboost>=2.0.3

# API and Data Collection
requests>=2.31.0
python-dateutil>=2.8.2

# Data Processing
scipy>=1.11.4
joblib>=1.3.2

# Visualization
matplotlib>=3.8.0
seaborn>=0.13.0
plotly>=5.18.0

# Configuration
pyyaml>=6.0.1

# Jupyter Notebooks
jupyter>=1.0.0
ipykernel>=6.27.0
notebook>=7.0.6

# Testing
pytest>=7.4.3
pytest-cov>=4.1.0

# Utilities
tqdm>=4.66.1
python-dotenv>=1.0.0

# Optional: Advanced ML
# lightgbm>=4.1.0
# catboost>=1.2
